Move ConnectRecordValueToMapTransformer of Elastisearch IT test in the connector
